backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.1
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.2